Make post images zoomable only if the parent is not an anchor
