+++ locales/fa.json
@@ -1,48 +1,48 @@
 {
-    "% min read": "% دقیقه",
-    "% posts": "% پست ها",
-    "1 min read": "۱ دقیقه",
-    "1 post": "یک پست",
-    "Among with": "<br>همراه با ",
-    "Among with no break line": "همراه با ",
-    "Apparently there are no posts at the moment, check again later.": "به نظر می رسد در حال حاضر پستی وجود ندارد. دوباره سر بزنید",
-    "Back to home": "بازگشت به خانه",
-    "Email cannot be blank.": "ایمیل نمی تواند خالی باشد",
-    "Featured": "برجسته",
-    "Get the latest posts delivered right to your inbox.": "جدیدترین پست ها را در ایمیل خود دریافت کنید",
-    "Go to the home page": "برو به صفحه اصلی",
-    "Great! Next, complete checkout for full access.": "عالیه! حالا برای دسترسی کامل وارسی رو تکمیل کنید",
-    "Great! You've successfully subscribed.": "عالیه! با موفقیت عضو شدید",
-    "JavaScript license information": "اطلاعات لایسنس جاوااسکریپت",
-    "More": "بیشتر",
-    "No posts": "پستی نیست",
-    "No posts found": "پستی پیدا نشد",
-    "No recent articles found :(": "مقاله تازه ای پیدا نشد :(",
-    "No results for your search, try something different.": "نتیجه ای برای جست و جوی شما پیدا نشد",
-    "No tags found :(": "تگی پیدا نشد :(",
-    "Now check your inbox and click the link to confirm your subscription.": "ایمیل خود را چک کنید و بر روی لینک فعال سازی کلیک کنید",
-    "Oops! There was an error sending the email, please try later.": "مشکلی در ارسال ایمیل به وجود آمد. دوباره امتحان کنید",
-    "Page": "صفحه",
-    "Page not found": "صفحه پیدا نشد",
-    "Please enter a valid email address": "لطفا یک ایمیل صحیح وارد کنید",
-    "Posted by": "پست شده توسط",
-    "Published with": "منتشر شده با",
-    "Recent articles": "مقالات جدید",
-    "Recommended for you": "پیشنهاد شده برای شما",
-    "Stay up to date! Get all the latest & greatest posts delivered straight to your inbox.": "به روز بمانید! جدیدترین و بهترین پست ها را در ایمیل خود دریافت کنید",
-    "Subscribe": "عضویت",
-    "Subscribe to": "عضو شدن به ",
-    "Subscribe to our newsletter": "در خبرنامه ما عضو شوید",
-    "Subscribed!": "عضو شدید!",
-    "Success! Your account is fully activated, you now have access to all content.": "موفق! اکانت شما فعال شد",
-    "Tags": "تگ ها",
-    "Toggle dark mode": "حالت تاریک",
-    "Type to search": "برای جست و جو بنویسید",
-    "Unfortunately the page you were looking for could not be found.": "متاسفانه صفحه ای که به دنبال آن می گردید پیدا نشد",
-    "Welcome back! You've successfully signed in.": "خوش آمدید! با موفقیت وارد شدید",
-    "You've successfully subscribed to": "شما با موفقیت عضو شدید به ",
-    "Your email address": "آدرس ایمیل شما",
-    "of": "از",
-    "with the email address": "با آدرس ایمیل",
-    "with this tag": "با این تگ"
-  }
+  "% min read": "% دقیقه",
+  "% posts": "% پست ها",
+  "1 min read": "۱ دقیقه",
+  "1 post": "یک پست",
+  "Among with": "<br>همراه با ",
+  "Among with no break line": "همراه با ",
+  "Apparently there are no posts at the moment, check again later.": "به نظر می رسد در حال حاضر پستی وجود ندارد. دوباره سر بزنید",
+  "Back to home": "بازگشت به خانه",
+  "Email cannot be blank.": "ایمیل نمی تواند خالی باشد",
+  "Featured": "برجسته",
+  "Get the latest posts delivered right to your inbox.": "جدیدترین پست ها را در ایمیل خود دریافت کنید",
+  "Go to the home page": "برو به صفحه اصلی",
+  "Great! Next, complete checkout for full access.": "عالیه! حالا برای دسترسی کامل وارسی رو تکمیل کنید",
+  "Great! You've successfully subscribed.": "عالیه! با موفقیت عضو شدید",
+  "JavaScript license information": "اطلاعات لایسنس جاوااسکریپت",
+  "More": "بیشتر",
+  "No posts": "پستی نیست",
+  "No posts found": "پستی پیدا نشد",
+  "No recent articles found :(": "مقاله تازه ای پیدا نشد :(",
+  "No results for your search, try something different.": "نتیجه ای برای جست و جوی شما پیدا نشد",
+  "No tags found :(": "تگی پیدا نشد :(",
+  "Now check your inbox and click the link to confirm your subscription.": "ایمیل خود را چک کنید و بر روی لینک فعال سازی کلیک کنید",
+  "Oops! There was an error sending the email, please try later.": "مشکلی در ارسال ایمیل به وجود آمد. دوباره امتحان کنید",
+  "Page": "صفحه",
+  "Page not found": "صفحه پیدا نشد",
+  "Please enter a valid email address": "لطفا یک ایمیل صحیح وارد کنید",
+  "Posted by": "پست شده توسط",
+  "Published with": "منتشر شده با",
+  "Recent articles": "مقالات جدید",
+  "Recommended for you": "پیشنهاد شده برای شما",
+  "Stay up to date! Get all the latest & greatest posts delivered straight to your inbox.": "به روز بمانید! جدیدترین و بهترین پست ها را در ایمیل خود دریافت کنید",
+  "Subscribe": "عضویت",
+  "Subscribe to": "عضو شدن به ",
+  "Subscribe to our newsletter": "در خبرنامه ما عضو شوید",
+  "Subscribed!": "عضو شدید!",
+  "Success! Your account is fully activated, you now have access to all content.": "موفق! اکانت شما فعال شد",
+  "Tags": "تگ ها",
+  "Toggle dark mode": "حالت تاریک",
+  "Type to search": "برای جست و جو بنویسید",
+  "Unfortunately the page you were looking for could not be found.": "متاسفانه صفحه ای که به دنبال آن می گردید پیدا نشد",
+  "Welcome back! You've successfully signed in.": "خوش آمدید! با موفقیت وارد شدید",
+  "You've successfully subscribed to": "شما با موفقیت عضو شدید به ",
+  "Your email address": "آدرس ایمیل شما",
+  "of": "از",
+  "with the email address": "با آدرس ایمیل",
+  "with this tag": "با این تگ"
+}
