+++ src/js/post.js
@@ -7,6 +7,8 @@ import shave from 'shave'
 import {
   isRTL,
   isMobile,
+  adjustImageGallery,
+  managePostImages,
   makeImagesZoomable
 } from './helpers'
 
@@ -114,18 +116,6 @@ $(document).ready(() => {
 
   fitvids('.js-post-content')
 
-  function adjustImageGallery() {
-    const images = document.querySelectorAll('.kg-gallery-image img')
-
-    for (var i = 0, len = images.length; i < len; i++) {
-      const container = images[i].closest('.kg-gallery-image')
-      const width = images[i].attributes.width.value
-      const height = images[i].attributes.height.value
-      const ratio = width / height
-      container.style.flex = `${ratio} 1 0%`
-    }
-  }
-
   adjustImageGallery()
   adjustShare(1000)
 
@@ -173,19 +163,7 @@ $(document).ready(() => {
     $commentsIframe.fadeIn('slow')
   })
 
-  $('.js-post-content').find('img').each(function() {
-    if (!$(this).closest('figure').hasClass('kg-bookmark-card')) {
-      $(this).addClass('js-zoomable')
-    }
-
-    const $figcaption = $(this).parent().find('figcaption')
-    if ($figcaption) {
-      $(this).attr('alt', $figcaption.text())
-    } else {
-      $(this).attr('alt', '')
-    }
-  })
-
+  managePostImages($)
   makeImagesZoomable($, mediumZoom)
 
   window.addEventListener('scroll', onScrolling, { passive: true })
